requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.28.64
pydantic==2.4.2
pymongo==4.5.0
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
pytest==7.4.2
pytest-asyncio==0.21.1
